@@ -13,11 +13,13 @@
   "dependencies": {
     "@babel/polyfill": "^7.11.5",
     "core-js": "^3.6.5",
-    "vue": "^2.6.12"
+    "vue": "^3.0.2"
   },
   "devDependencies": {
     "@babel/core": "^7.11.6",
     "@babel/preset-env": "^7.11.5",
+    "@vue/babel-plugin-jsx": "^1.0.0-rc.3",
+    "@vue/compiler-sfc": "^3.0.2",
     "autoprefixer": "^10.0.1",
     "babel-loader": "^8.1.0",
     "clean-webpack-plugin": "^3.0.0",
@@ -30,12 +32,10 @@
     "postcss-preset-env": "^6.7.0",
     "sass-loader": "^10.0.2",
     "style-loader": "^1.2.1",
+    "vue-loader": "^16.0.0-beta.9",
     "webpack": "^4.44.2",
     "webpack-cli": "^3.3.12",
-    "webpack-merge": "^5.1.4",
-    "vue-loader": "^15.9.3",
-    "vue-style-loader": "^4.1.2",
-    "vue-template-compiler": "^2.6.12"
+    "webpack-merge": "^5.1.4"
   },
   "browserslist": [
     "> 1%",
