@@ -1,9 +1,4 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
 
-const root = document.createElement('div')
-document.body.appendChild(root)
-
-new Vue({
-    render:(h) => h(App)
-}).$mount(root)
+createApp(App).mount('#app')
