@@ -2,7 +2,8 @@ const path = require("path");
 
 const webpack = require("webpack");
 const { CleanWebpackPlugin } = require("clean-webpack-plugin");
-const VueLoaderPlugin = require("vue-loader/lib/plugin");
+
+const { VueLoaderPlugin } = require("vue-loader");
 
 function resolve(dir) {
   return path.join(__dirname, dir);
